Horizontal pattern scroll and misc. fixes.
* private.h : (Line 457)
It turns out to be a bad idea to make xpos ULONG to begin with. Now it is LONG and all the typecasts to prevent compiler warnings are also removed throughout the code.

* GetSetAttrs.c : mSet() (Lines 381, 388 and 416 )
With the changes made, now we (vertically) scroll all the contents of the gadget, not only the used part of it as it was before. This is to prevent the glitches at the bottom of the gadget as scrolling when gadget height is not a multiple of fontheight and is dependent on the changes made in the DumpText() function.

(Line 459)
A change in visual_y is no longer a case for longestline re-calculation (it was only suitable for the deprecated SliderBehaviour). This is why it is commented out. 
(Jens please remove it from code if you please.)

* Dispatcher.c : mDraw() (Line 611)
With the changes made, DumpText() now can intelligently clear the very last part of the gadget content. This section is no longer a requirement